ci: builder: support building with --cap-add=SYS_PTRACE
In their infinite wisdom, the docker devs decided in
https://github.com/moby/moby/issues/1916 that adding support
for --cap-add=SYS_PTRACE for `docker build` is not a priority
or that they just don't want to do that.

So we have to work around this limitation by creating a
temporary base image, then run the install script with
`docker run`. Next, the resulting container is converted to
an image with `docker commit`. Finally we clean up the mess
we have made.